drivers: gpio: pca_series: Fix integer overflow
The Coverity static analysis tool identified a potential integer overflow
in the gpio_pca_series_pin_interrupt_configure_extended function.
The 'pin' variable, a uint8_t, is left-shifted before being used in a
64-bit calculation. If 'pin' is close to its maximum value, the
intermediate calculation could overflow.

This commit fixes the issue by casting 'pin' to a uint32_t before the
left shift, ensuring the calculation is performed with a wider type and
preventing the potential overflow.

Fixes #81963

+++ drivers/gpio/gpio_pca_series.c
@@ -1358,7 +1358,7 @@ static int gpio_pca_series_pin_interrupt_configure_extended(
 	uint64_t int_edge;
 	uint32_t int_mask, input_latch;
 	int ret = 0;
-	uint32_t edge_cfg_shift = pin << 1U;
+	uint32_t edge_cfg_shift = (uint32_t)pin << 1U;
 	uint64_t edge_cfg_mask = 0b11 << edge_cfg_shift;
 
 	if (cfg->gpio_int.port == NULL) {
